ml-backend-flask/requirements.txt
addict==2.4.0
async-timeout==4.0.3
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
defusedxml==0.7.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
fonttools==4.45.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jstyleson==0.0.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.1
numpy==1.26.2
opencv-python==4.8.1.78
openvino==2023.2.0
openvino-dev==2023.2.0
openvino-telemetry==2023.2.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rq==1.15.1
scikit-learn==1.3.2
scipy==1.11.4
simple-websocket==1.0.0
six==1.16.0
texttable==1.7.0
threadpoolctl==3.2.0
tqdm==4.66.1
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
